+++ src/DocSharp.Docx/DocxToHtml/MapBorderSpacing.cs
@@ -1,0 +1,8 @@
+namespace DocSharp.Docx;
+
+internal enum MapBorderSpacing
+{
+    None,
+    Padding,
+    Margin
+}